Optimal Soln for Climbing Stairs with O(N) time complexity and O(1) space Complexity

+++ 13-1D_Dynamic_Programming/Climbing-Stairs.cpp
@@ -1,10 +1,10 @@
 class Solution {
 public:
     int climbStairs(int n) {
-        unsigned long long int i=1;
+        int i=1;
         long int j=1;
-        unsigned long long int k=1;
-        unsigned long long int h=0;
+        int k=1;
+        int h=0;
         if (n==1){
             return 1;
         }
